moodify_site/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.20
django-markdown-deux==1.0.5
idna==2.8
markdown2==2.3.7
music-recommender==1.0.5
oauthlib==3.0.1
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
spotipy==2.4.4
tweepy==3.7.0
urllib3==1.24.1
watson-developer-cloud==2.8.1
websocket-client==0.48.0
